Replace Floodgate with Geyser at packet level
It is possible to use Geyser without Floodgate by configuring Geyser to
use auth-type= 'online' or 'offline'. In that scenario, floodgateService
may be unavailable, however, Geyser players are still unable to reply to
premium Java login requests. Moreover, the FloodgateApi is unreliable
at this point of the auth process, so actual Floodgate players may not
be part of its "online-players" list (see #603).

FloodgateService is kept in JoinManagement, because a prefix check is
performed for cracked Java players. However, these prefixes are
Floodgate specific and are not available on Geyser only servers.
Therefore, it is necessary to check if Floodgate is installed.

It is also possible to run Floodgate without Geyser on Bukkit servers,
if Geyser is installed on a Bungee proxy. In that case, FastLogin will
also use BungeeCord, which means that the code at JoinManagement will
not get executed by the Bukkit instance. Therefore, it is impossible
for a Floodgate player to reach JoinManagement on a Floodgate only
server.

This commit removes name conflict checking.