Make "Floodgate" nullable in StoredProfile
The Floodgate field is null in the database, if the player has not been migrated to the new database format.
It's a bit cleaner this way than storing the isFloodgateMigrated in a separate field in StoredProfile. One less variable in the already crowded constructor.
While this is not memory efficient, this is only temporary, as per https://github.com/games647/FastLogin/pull/709#issuecomment-1207829307 (last sentence).